feat(dataframe): make DataFrame::builder available without rayon (serial fill) (#1055)
Relocate `RDataFrameBuilder` and its helper types (`RangeFiller`, `ColPtr`,
`ColumnKind`, `ColumnReg`) out of the rayon-gated
`miniextendr-api/src/optionals/rayon_bridge.rs` into a new ungated module
`miniextendr-api/src/dataframe_builder.rs`. The only rayon-specific code — the
flattened parallel fill pass in `build_sexp` plus `rayon::current_num_threads()`
— is now gated inside `build`; without `rayon` each column is filled in one
full-range serial pass, producing an identical `data.frame`.

`DataFrame::builder` loses its `#[cfg(feature = "rayon")]` and now returns
`crate::dataframe_builder::RDataFrameBuilder`. The documented
`crate::rayon_bridge::RDataFrameBuilder` path still resolves via a back-compat
re-export under `rayon`. Add `gc_stress_dataframe_builder_serial()` exercising
the serial path (rpkg compiles with rayon off by default) and a feature-agnostic
unit test proving the builder surface is callable without rayon.
