+++ rpkg/NAMESPACE
@@ -905,6 +905,7 @@ export(gc_stress_builder_grow_schema)
 export(gc_stress_builder_with_schema)
 export(gc_stress_condition_data)
 export(gc_stress_cow_subslice_roundtrip)
+export(gc_stress_dataframe_builder_serial)
 export(gc_stress_dataframe_map)
 export(gc_stress_dataframe_nested_enum)
 export(gc_stress_dataframe_par_reader)
