+++ rpkg/man/gc_stress_fixtures.Rd
@@ -41,6 +41,7 @@
 \alias{gc_stress_dispatch_to_dataframes}
 \alias{gc_stress_dataframe_to_vec_nested}
 \alias{gc_stress_arrow_sliced_recordbatch}
+\alias{gc_stress_dataframe_builder_serial}
 \alias{gc_stress_externalptr_collect_list}
 \alias{gc_stress_result_to_dataframe_auto}
 \alias{gc_stress_result_to_dataframe_collated}
@@ -126,6 +127,8 @@ Generated by miniextendr from Rust fn \code{gc_stress_dispatch_to_dataframes}
 Generated by miniextendr from Rust fn \code{gc_stress_dataframe_to_vec_nested}
 
 Generated by miniextendr from Rust fn \code{gc_stress_arrow_sliced_recordbatch}
+
+Generated by miniextendr from Rust fn \code{gc_stress_dataframe_builder_serial}
 
 Generated by miniextendr from Rust fn \code{gc_stress_externalptr_collect_list}
 
@@ -216,6 +219,8 @@ gc_stress_dataframe_to_vec_nested()
 
 gc_stress_arrow_sliced_recordbatch()
 
+gc_stress_dataframe_builder_serial()
+
 gc_stress_externalptr_collect_list()
 
 gc_stress_result_to_dataframe_auto()
@@ -233,6 +238,8 @@ gc_stress_result_to_dataframe_split_sentinel()
 Production helper for the \verb{&str} zero-copy argument-borrow path (#664)
 
 Convert an R vector to an ALTREP-backed vector by materializing then re-wrapping
+
+gc stress fixtures
 
 gc stress fixtures
